+++ src/main/skripte/db-versioning-templates/uebergreifend/db-update-x.y.z_1/update-db-schema.sh
@@ -1,27 +1,31 @@
 #!/bin/bash
-###
+# #%L
+# plis-persistence
+# %%
+# 
+# %%
 # See the NOTICE file distributed with this work for additional
 # information regarding copyright ownership.
 # The Federal Office of Administration (Bundesverwaltungsamt, BVA)
 # licenses this file to you under the Apache License, Version 2.0 (the
 # License). You may not use this file except in compliance with the
 # License. You may obtain a copy of the License at
-#
+# 
 #     http://www.apache.org/licenses/LICENSE-2.0
-#
+# 
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
 # implied. See the License for the specific language governing
 # permissions and limitations under the License.
-###
+# #L%
 
 ########################################################################
 # Aktualisiert die Datenbank für das System in der Umgebung.
 #######################################################################
 LogFile=logs/ausgabe.log
 
-sqlplus -S /nolog @00_update-main.sql 01_environment.sql ${LogFile}
+sqlplus -S /nolog @00_update-main.sql 91_environment.sql ${LogFile}
 echo ''
 
 AnzahlFehler=$(egrep -c "SP2\-[0-9]{4}\:|ORA\-[0-9]{5}\:" ${LogFile})